requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
galois==0.4.4
kiwisolver==1.4.9
llvmlite==0.44.0
matplotlib==3.10.8
numba==0.61.2
numpy==2.2.6
packaging==26.0
pillow==12.1.1
pyadi-iio==0.0.20
pylibiio==0.25
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scipy==1.17.1
sdr==0.0.29
six==1.17.0
typing_extensions==4.15.0
